@@ -4,27 +4,28 @@
   "private": true,
   "homepage": "/ColorTransferLab/",
   "dependencies": {
-    "@react-three/cannon": "^4.6.1",
-    "@react-three/drei": "^8.11.0",
-    "@react-three/fiber": "^7.0.24",
-    "@react-three/xr": "^3.5.0",
-    "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/user-event": "^13.5.0",
-    "@babel/plugin-proposal-private-property-in-object":"7.21.0",
+    "@babel/plugin-proposal-private-property-in-object": "7.21.11",
+    "@mkkellogg/gaussian-splats-3d": "^0.4.5",
+    "@react-three/cannon": "^6.6.0",
+    "@react-three/drei": "^9.114.0",
+    "@react-three/fiber": "^8.17.9",
+    "@testing-library/jest-dom": "^6.5.0",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
     "file-system": "^2.2.2",
-    "history": "^5.2.0",
+    "history": "^5.3.0",
     "install": "^0.13.0",
-    "jquery": "^3.7.0",
-    "mini-css-extract-plugin": "^2.4.5",
-    "node-sass": "^7.0.1",
-    "npm": "^8.10.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1",
-    "react-scripts": "^5.0.0",
-    "three": "^0.138.0",
-    "web-vitals": "^2.1.3"
+    "jquery": "^3.7.1",
+    "mini-css-extract-plugin": "^2.9.1",
+    "npm": "^10.9.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.2",
+    "react-scripts": "^5.0.1",
+    "react-zoom-pan-pinch": "^3.6.1",
+    "sass": "^1.79.4",
+    "three": "^0.161.0",
+    "web-vitals": "^4.2.3"
   },
   "scripts": {
     "start": "BROWSER='none' react-scripts start",
@@ -51,6 +52,6 @@
     ]
   },
   "devDependencies": {
-    "@types/jquery": "^3.5.16"
+    "@types/jquery": "^3.5.31"
   }
 }
